Revert "mudando no search_node: node_id_2 de size_t para int(para que seja possível passar -1 para ele)"
This reverts commit 764f4cb702cc4c7a25d6b8af121af190779a2cc3.

+++ include/Graph.hpp
@@ -36,7 +36,7 @@ private:
 
 
     //* Funções Auxiliares
-    void search_nodes(Node *node_1, size_t node_id_1, Node *node_2 = nullptr, int node_id_2 = -1);
+    void search_nodes(Node *node_1, size_t node_id_1, Node *node_2 = nullptr, size_t node_id_2 = -1);
     void aux_remove_edge(Node *node_1, size_t node_id_2);
 };
